pyproject.toml
[tool.poetry]
name = "coupon-core"
version = "0.1.0"
description = "Coupon Scraper is a web-based application designed to help users save money by discovering and collecting discount codes and coupons from various online sources. The application leverages advanced web scraping technologies and APIs to curate a personalized collection of active deals, offering users a seamless way to find and use discounts in real-time."
authors = ["Ireoluwa Fakeye <faksiret@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
django = ">=5.1.4,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
djangorestframework-simplejwt = ">=5.3.1,<6.0.0"
celery = ">=5.4.0,<6.0.0"
redis = ">=5.2.1,<6.0.0"
django-storages = {extras = ["boto3"], version = ">=1.14.4,<2.0.0"}
python-dotenv = ">=1.0.1,<2.0.0"
requests = ">=2.32.3,<3.0.0"
daphne = ">=4.1.2,<5.0.0"
channels = ">=4.2.0,<5.0.0"
django-cors-headers = ">=4.6.0,<5.0.0"
boto3 = ">=1.35.85,<2.0.0"
channels-redis = ">=4.2.1,<5.0.0"
types-redis = ">=4.6.0.20241004,<5.0.0"
bs4 = "^0.0.2"
beautifulsoup4 = "^4.12.3"
tenacity = "^9.0.0"
pyrate-limiter = "^3.7.0"
pytest = "^8.3.4"
geopy = "^2.4.1"
pinecone-client = "^5.0.1"
transformers = "^4.48.1"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
exclude = [".venv", "venv", "migrations"]


[tool.black]
exclude = '''
/(
    \.venv
    |venv
    |migrations
    |docker
    |Scripts
    |poetry\.lock
)
'''
line-length = 88
